matecomponent: Remove API to build matecomponent applet as shared library
Yes, it's bad, it breaks API and ABI. In a stable branch. But it's the
matecomponent compatibility module. And it's a rather important fix.

We can't have matecomponent applets as shared libraries since both matecomponent
applets and dbus applets are sharing the MatePanelApplet type name, and they
will conflict in the same process. Since we know we want to allow dbus
in-process applets, it means we have to disallow matecomponent in-process
applets. Hence this change.