bonobo: Remove API to build bonobo applet as shared library Yes, it's bad, it breaks API and ABI. In a stable branch. But it's the bonobo compatibility module. And it's a rather important fix.
We can't have bonobo applets as shared libraries since both bonobo
applets and dbus applets are sharing the PanelApplet type name, and they
will conflict in the same process. Since we know we want to allow dbus
in-process applets, it means we have to disallow bonobo in-process
applets. Hence this change.

	from http://git.gnome.org/browse/gnome-panel/commit/?h=gnome-2-32&id=cc7f3e7cac8dd9da5e8e48eac6ef9e992c557812