Fix rendering of user backgrounds in gtk3.18+
User set panel backgrounds other than colors with an alpha value fail on restarting the panel using gtk3.18 or 3.19. This makes picture and fully opaque color backgrounds set by the user work again and for some reason does not break transparent color backgrounds either. Tested with gtk3.19.5, should work with gtk3.18 as this bug was exactly the same but needs testing with it.