gtk3.20: custom background work on main panel
This makes custom backgrounds show on the panel, the menu, the launcher applets
but not yet the tray, the clock button, the trash, or most applets from mate-applets.
Sometimes panel BG not drawn after theme change until panel restarted.
Still, this is a clue-will keep working.
Adding gdk_window_ensure_native (window) was taken from gnome-panel but caused the
default background not to show if the system BG is selected and does not explicitly
set the theme. Also using this brings back the requirement to restart the panel to
fully apply a system BG but progress is being made here.