GTK 3.20 force custom BG drawing on applets
Use gdk_window_ensure_native to applets when and only when a custom background is selected.
Doing so when the system bg is selected will turn applets clear,
so will not doing so when a custom background is selected.
Keep all this behind if # GTK_CHECK_VERSION (3, 19, 0) so as not to screw up anything
in older GTK versions.