GTK 3.20-Prevent transparent panel on BG change
The mate-custom-panel-background style class persists on the panel widget when resetting the panel BG to the system theme. It will overrride the panel system theme and turn the panel transparent with the new applet changes for some reason, so don't run panel_background_apply_css on GTK 3.20 builds. It's not needed to override t.he theme background for a custom background with the new code and GTK 3.20, and mate-panel-menu-bar is already present