Add (optional) support for menulibre. https://smdavis.us/projects/menulibre/
This patch repurposes the existing fallback support for the now obsolete
matemenu-simple-editor to add (optional) support for menulibre. The logic is
simple, if menulibre is installed use it in preference to mozo. If menulibre
is not available, fallback to mozo.

The rationale is this; if you've installed menulibre you want to use it
in preference to mozo. Menulibre offers a superset of menu editting
capabilities compared to mozo and is growing in popularity. This change will
allow distro maintainers to chose which menu editor they want
to offer.