status-notifier-watcher: Allow re-registration of registered items
The SNI specification doesn't explicitly state what should happen when
trying to register again an already registered item, so it would seem
only logical to forbid it.  Unfortunately, libappindicator tries
re-registering pretty often, and even falls back to System Tray if it
fails.  So in practice we need to be forgiving and pretend it's OK,
or we get a duplicate System Tray item for this one.

Fixes #539.