Run dialog: Fix wrong history order
Initially the org.mate.panel.general.history-mate-run key was saved with
the most recent command last. In combination with wrong code this led to
scrambling the history.

This commit changes the history storage in a way that most recent
elements are stored first. The code adapts to this change by changing
only one function to prepend instead of append in the list store. (Any
old histories might therefore be wrong for a short time, but that's
probably worth it to fix this bug.)