mate-panel: Dynamically apply a wmclass to control panel shadow
In order to further facilitate mate-desktop/marco#327, this change
will dynamically set the wmclass on each panel start to one respecting
alpha settings.

That is to say, in only the themed (PANEL_BACK_NONE) mode, do we want
to actually have a shadow on mate-panel. To do this, we'll maintain
the "mate-panel" wmclass, which will be dealt with by Marco on map.
For all other cases we'll change the wmclass to "mate-panel-alpha"
to indicate that we're an RGBA window with non theme styling, and
as such do not expect any shadowing.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>