window-list: Show window previews on hover
When hovering on taskbar entries, display a popup with a thumbnail
of the associated window. Popup goes away when leaving the button.

If a window has not been mapped since the applet started, it will
not display a thumbnail. Once mapped, however, it will use its
latest available thumbnail as preview, even when minimized.

Thumbnails can be turned on or off, and the size can be changed in
the applet preferences dialog.