Avoid double a11y description on panel toplevels
Do not set an extra a11y name and description on an internal child of the
panel, because its toplevel has most of it already through setting the GTK
window title.  Instead, explicitly set the a11y name and description on the
toplevel directly, possibly overriding GTK's value implicitly set via
gtk_window_set_title(), but which would be the same anyway.

This prevents e.g. a screen reader vocalizing the panel name twice when
entering it, once for the toplevel and once for the internal child.
This also avoids announcing the panel again when moving from one of the hide
buttons to one of the applets or launcher.