Fix creating several launchers at once
When creating a new launcher, we try and find an available object ID by
browsing the existing ones.  This however fails to take into account
objects that were created but that didn't make it to the backend yet,
leading to likely re-using the same ID when creating several objects
roughly at the same time.  This happens e.g. when dropping several URIs
together.