Ask the user what to do when receiving an image drop
Instead of unconditionally setting the panel's background when an image
file URI gets dropped on a panel, ask the user whether to create a
launcher for that URI (like for non-image URIs) or to use the image as
the panel's background.

A number of users have seen their panel's background changed
unexpectedly due to unwanted drop of an image file over the panel,
sometimes seeing this as a panel's bug.

Also, this behavior is inconsistent with dropping a URI pointing to any
other type of file, where it would create a launcher for it.

Consequently, and as setting the panel's background doesn't seem like a
so common task it ought to be super fast, ask the user what to do when
receiving an image URI drop.