Add "center-stick" capability applets on the panel
Conventionally, the applets on a MATE Panel are positioned relative to the
left edge of the panel (if the panel is horizontal) or the top edge (if the
panel is vertical).  There has also been some (buggy) support for positioning
of applets relative to the right (or bottom) edge of the panel, so that
applets on the right side of the panel will stay on the right side even if
the user changes screen resolutions or if the panel changes size for some
other reason.

However, many users want to also place applets at or near the center of their
panel(s).  There is no such conventional support for positioning applets
relative to the center of the panel, so users have positioned applets near
the center of the panel -- but the position recorded is relative to the left
side of the panel.  As such, the applets will almost certainly shift over
to the left or right slightly if the panel is ever resized, and the user will
have to reposition all those centered applets yet again.  This is especially
frustrating if the user switches monitors on a regular basis!

This patch radically revamps the MATE Panel's positioning framework, and
deprecates the original "right-stick" feature.  To replace the right-stick
feature, this patch instead associates an "edge relativity" setting with
each and every panel applet:  An applet can be relative to the start (left/
top), end (right/bottom), or center of the panel.  This setting can be
changed using DConf/GSettings, using a custom panel layout file, or even
by simply dragging the applet to the appropriate place on the panel.
(Conventionally, applets are not even right-stuck automatically even when the
user drags the applet over to the far right of the panel!)  As a bonus, when
the user drags an applet across the center of the panel, the applet will
temporarily "stick" to the very center of the panel, to allow the user to
very precisely align any applet they wish.