Clock Applet: Update date/time and weather after resume from sleep mode
Before this commit, the Clock Applet was unaware of when the system
resumed from sleep mode (e.g. suspend or hibernate).  This meant that,
if the user suspended the system and then woke it up some time later,
the clock might not show the real current time, but instead would continue
to show the time at which the system was suspended for up to a minute
after the system woke up.  This could be confusing to the user if they
peered up at the clock immediately after, say, opening their laptop lid.

This commit connects to ConsoleKit or (if available) systemd-logind via
D-Bus, and listens for a signal that indicates when the system wakes up
from sleep mode.  When the system does wake up, the Clock Applet updates
the date/time immediately, and also updates all the weather/temperature
readings if possible.

Note that if logind is missing, ConsoleKit2 is required for this feature
to work; the original ConsoleKit lacked the signal required to implement
this feature.  This should not be a problem since it is unlikely that
anyone would even consider using the original ConsoleKit in this day and
age.