wnck-pager: Fix workspace switcher aspect ratio
This is a dirty hack, but it works both in-process and out-of-process,
and doesn't break window list on non-expanded panels.

A better solution would be to actually fix the panel's layout code, but
nobody seems to have enough time and experience with this to actually
do this at the moment.

Based off an idea and code from Alberts Muktupāvels[^1], thanks.

Fixes #1230.

[^1]: https://github.com/mate-desktop/mate-panel/issues/1230#issuecomment-1046235088