Clock Applet: Add scrollbar to Locations lists when the lists get too long
Before this commit, the Clock Applet's popup window would expand off the
screen if the user added more than a few locations to the Locations list.
In addition, the list of locations in the Clock Applet Preferences could get
extremely tall as well, and could even push the bottom part of the
Preferences dialog off the screen.

This commit adds (vertical) scrollbars to both widgets.