clock: Store calendar client in ClockData (#1516)
* clock: Store calendar client in ClockData

This simplifies the evolution calendar client architecture and makes the
popup show up faster, since we don't need to create a new one every time
we open it.

* clock: Fix calendar data refresh when opening popup

The calendar popup was not showing recently added Evolution
events/tasks. Added a refresh timeout call every time the calendar popup
opens. That way the popup opens immediately and the data is refresh
after a small delay.